move tx LS
